+++ src/main/scala/dxWDL/util/InstanceTypeDB.scala
@@ -53,7 +53,7 @@ case class DxInstanceType(name: String,
                           diskGB: Int,
                           cpu: Int,
                           price: Float,
-                          os: Vector[(String, String)]) extends Ordered[DxInstanceType] {
+                          os: Vector[(String, String)]) {
     // Does this instance satisfy the requirements?
     def satisfies(memReq: Option[Int],
                   diskReq: Option[Int],
@@ -73,47 +73,55 @@ case class DxInstanceType(name: String,
         return true
     }
 
-    // Comparison function. Returns true iff [this] comes before
-    // [that] in the ordering.
-    //
-    // If the hourly price list per instance in available, we sort by
-    // price. If we do not have permissions for pricing information,
-    // we compare by resources sizes. For example, if A has more
-    // memory, disk space, and cores than B, then B < A. We round down
-    // memory and disk sizes, to make the comparison insensitive to
-    // minor differences.
-    def lteq(that: DxInstanceType) : Boolean = {
-        compare(that) <= 0
-    }
-
-    def compare(that: DxInstanceType) : Int = {
+    def compareByPrice(that: DxInstanceType) : Int = {
         // compare by price
         if (this.price < that.price)
             return -1
         if (this.price > that.price)
             return 1
+        return 0
+    }
 
-        // Prices are the same, compare based on resource sizes.
-        // We add some fuzziness to the comparison, because the instance
-        // types don't have the exact memory and disk space that you would
-        // expect. For example, mem1_ssd1_x2 has less disk space than mem2_ssd1_x2.
+    // Compare based on resource sizes. This is a partial ordering.
+    //
+    // For example, if A has more memory, disk space, and cores than
+    // B, then B < A. We round down memory and disk sizes, to make the
+    // comparison insensitive to minor differences.
+    //
+    // We add some fuzziness to the comparison, because the instance
+    // types don't have the exact memory and disk space that you would
+    // expect. For example, mem1_ssd1_x2 has less disk space than mem2_ssd1_x2.
+    def compareByResources(that : DxInstanceType) : Int = {
         val memDelta = (this.memoryMB/1024) - (that.memoryMB/1024)
         val diskDelta = (this.diskGB/16) - (that.diskGB/16)
         val cpuDelta = this.cpu - that.cpu
 
-        if (memDelta == 0 &&
-                diskDelta == 0 &&
-                cpuDelta == 0)
-            return 0
-        if (memDelta <= 0 &&
-                diskDelta <= 0 &&
-                cpuDelta <= 0)
-            return -1
-        if (memDelta >= 0 &&
-                diskDelta >= 0 &&
-                cpuDelta >= 0)
-            return 1
-        return 0;
+        val retval =
+            if (memDelta == 0 && diskDelta == 0 && cpuDelta == 0)
+                0
+            else if (memDelta <= 0 && diskDelta <= 0 && cpuDelta <= 0)
+                -1
+            else if (memDelta >= 0 &&  diskDelta >= 0 && cpuDelta >= 0)
+                1
+            else
+                // instances cannot be directly compared.
+                0
+        //System.out.println(s"compareByResource ${this.name} ${that.name} retval=${retval}")
+        retval
+    }
+
+    // v2 instances are always better than v1 instances
+    def compareByType(that: DxInstanceType) : Int = {
+        //System.out.println(s"compareByType ${this.name} ${that.name}")
+        def typeVersion(name: String) =
+            if (name contains "_v2") "v2"
+            else "v1"
+
+        (typeVersion(this.name), typeVersion(that.name)) match {
+            case ("v1", "v2") => -1
+            case ("v2", "v1") => 1
+            case (_, _) => 0
+        }
     }
 }
 
@@ -123,7 +131,25 @@ object DxInstanceType extends DefaultJsonProtocol {
 }
 
 
-case class InstanceTypeDB(instances: Vector[DxInstanceType]) {
+case class InstanceTypeDB(pricingAvailable : Boolean,
+                          instances: Vector[DxInstanceType]) {
+
+    // if prices are available, choose the cheapest instance. Otherwise,
+    // choose one with minimal resources.
+    private def lteq(x : DxInstanceType, y : DxInstanceType) : Boolean = {
+        val costDiff =
+            if (pricingAvailable) {
+                x.compareByPrice(y)
+            } else {
+                x.compareByResources(y)
+            }
+        if (costDiff != 0)
+            return costDiff <= 0
+
+        // cost is the same, compare by instance type. Better is HIGHER
+        x.compareByType(y) >= 0
+    }
+
     // Calculate the dx instance type that fits best, based on
     // runtime specifications.
     //
@@ -142,13 +168,12 @@ case class InstanceTypeDB(instances: Vector[DxInstanceType]) {
         val sufficient: Vector[DxInstanceType] =
             instances.filter(x => x.satisfies(memoryMB, diskGB, cpu))
         if (sufficient.length == 0)
-            throw new Exception(s"No instances found that match the requirements (memory=$memoryMB, diskGB=$diskGB, cpu=$cpu")
-
-        // if prices are available, choose the cheapest instance. Otherwise,
-        // choose one with minimal resources.
+            throw new Exception(s"""|No instances found that match the requirements
+                                    |memory=$memoryMB, diskGB=$diskGB, cpu=$cpu"""
+                                    .stripMargin.replaceAll("/", " "))
         val initialGuess = sufficient.head
         val bestInstance = sufficient.tail.foldLeft(initialGuess){ case (bestSoFar,x) =>
-            if (x.lteq(bestSoFar)) x
+            if (lteq(x, bestSoFar)) x
             else bestSoFar
         }
         bestInstance.name
@@ -174,13 +199,13 @@ case class InstanceTypeDB(instances: Vector[DxInstanceType]) {
         if (iTypes.isEmpty)
             throw new Exception("empty list")
         iTypes.tail.foldLeft(iTypes.head) {
-            case (cheapest, elem) =>
-                if (elem.lteq(cheapest)) elem
-                else cheapest
+            case (best, elem) =>
+                if (lteq(elem, best)) elem
+                else best
         }
     }
 
-    // An fast but cheap instance type.
+    // A fast but cheap instance type.
     //
     def defaultInstanceType : String = {
         val iType = instances.find(x => x.name == InstanceTypeDB.DEFAULT_INSTANCE_TYPE) match {
@@ -190,7 +215,28 @@ case class InstanceTypeDB(instances: Vector[DxInstanceType]) {
         iType.name
     }
 
-    def apply(iType: InstanceTypeReq) = {
+  // check if instance type A is smaller or equal in requirements to
+    // instance type B
+    def lteqByResources(iTypeA: String,
+                        iTypeB: String) : Boolean = {
+        val ax = instances.find(_.name == iTypeA)
+        val bx = instances.find(_.name == iTypeB)
+
+        (ax,bx) match {
+            case (Some(a), Some(b)) =>
+                // All the resources for instance A should be less than or equal
+                // to the resoruces for B.
+                (a.memoryMB <= b.memoryMB &&
+                     a.diskGB <= b.diskGB &&
+                     a.cpu <= b.cpu)
+            case (_,_) =>
+                // At least one of the instances is not in the database.
+                // We can't compare them.
+                false
+        }
+    }
+
+    def apply(iType: InstanceTypeReq) : String = {
         iType.dxInstanceType match {
             case None =>
                 choose3Attr(iType.memoryMB, iType.diskGB, iType.cpu)
@@ -211,34 +257,13 @@ case class InstanceTypeDB(instances: Vector[DxInstanceType]) {
         }
         sortediTypes.toJson.prettyPrint
     }
-
-    // check if instance type A is smaller or equal in requirements to
-    // instance type B
-    def lteqByResources(iTypeA: String,
-                        iTypeB: String) : Boolean = {
-        val ax = instances.find(_.name == iTypeA)
-        val bx = instances.find(_.name == iTypeB)
-
-        (ax,bx) match {
-            case (Some(a), Some(b)) =>
-                // All the resources for instance A should be less than or equal
-                // to the resoruces for B.
-                (a.memoryMB <= b.memoryMB &&
-                     a.diskGB <= b.diskGB &&
-                     a.cpu <= b.cpu)
-            case (_,_) =>
-                // At least one of the instances is not in the database.
-                // We can't compare them.
-                false
-        }
-    }
 }
 
 object InstanceTypeDB extends DefaultJsonProtocol {
     val DEFAULT_INSTANCE_TYPE = "mem1_ssd1_x4"
 
     // support automatic conversion to/from JsValue
-    implicit val instanceTypeDBFormat = jsonFormat1(InstanceTypeDB.apply)
+    implicit val instanceTypeDBFormat = jsonFormat2(InstanceTypeDB.apply)
 
     // Currently, we support only constants.
     def parse(dxInstanceType: Option[WomValue],
@@ -484,7 +509,7 @@ object InstanceTypeDB extends DefaultJsonProtocol {
             .filter{ case (iName,traits) => instanceCriteria(traits) }
             .map{ case (_,traits) => traits}
             .toVector
-        InstanceTypeDB(iTypes)
+        InstanceTypeDB(false, iTypes)
     }
 
     private def queryWithPrices(dxProject: DXProject) : InstanceTypeDB = {
@@ -502,7 +527,7 @@ object InstanceTypeDB extends DefaultJsonProtocol {
 
         // filter out instances that we cannot use
         val iTypes: Vector[DxInstanceType] = availableInstanceTypes.filter(instanceCriteria)
-        InstanceTypeDB(iTypes)
+        InstanceTypeDB(true, iTypes)
     }
 
     def query(dxProject: DXProject, verbose:Verbose) : InstanceTypeDB = {
@@ -535,149 +560,18 @@ object InstanceTypeDB extends DefaultJsonProtocol {
     def opaquePrices(db: InstanceTypeDB) : InstanceTypeDB = {
         if (db.instances.isEmpty)
             return db
-        // check if all the prices are zero, there is nothing to do then
-        val all_prices_are_zero = db.instances.forall(it => it.price == 0)
-        if (all_prices_are_zero)
+        // check if there is no pricing information
+        if (!db.pricingAvailable)
             return db
-        // sorted the prices from low to high, and then replace
+
+        // sort the prices from low to high, and then replace
         // with rank.
         var crnt_price = 0
-        val opaque = db.instances.sorted.map{ it =>
+        val sortedInstances = db.instances.sortWith(_.price < _.price)
+        val opaque = sortedInstances.map{ it =>
             crnt_price += 1
             it.copy(price = crnt_price)
         }
-        InstanceTypeDB(opaque)
-    }
-
-    // The original list is at:
-    // https://github.com/dnanexus/nucleus/blob/master/node_modules/instance_types/aws_instance_types.json
-    //
-    // The g2,i2,x1 instances have been removed, because they are not
-    // enabled for customers by default.  In addition, the PV (Paravirtual)
-    // instances have been removed, because they work only on Ubuntu
-    // 12.04.
-    //
-    // Removed the ssd2 instances, because they actually use EBS storage. A better
-    // solution would be asking the platform for the available instances.
-    private val instanceList : String = """{
-        "mem2_ssd1_x2":       {"internalName": "m3.large",                          "traits": {"numCores":   2, "totalMemoryMB":    7225, "ephemeralStorageGB":   27}},
-        "mem2_ssd1_x4":       {"internalName": "m3.xlarge",                         "traits": {"numCores":   4, "totalMemoryMB":   14785, "ephemeralStorageGB":   72}},
-        "mem2_ssd1_x8":       {"internalName": "m3.2xlarge",                        "traits": {"numCores":   8, "totalMemoryMB":   29905, "ephemeralStorageGB":  147}},
-        "mem1_ssd1_x2":       {"internalName": "c3.large",                          "traits": {"numCores":   2, "totalMemoryMB":    3766, "ephemeralStorageGB":   28}},
-        "mem1_ssd1_x4":       {"internalName": "c3.xlarge",                         "traits": {"numCores":   4, "totalMemoryMB":    7225, "ephemeralStorageGB":   77}},
-        "mem1_ssd1_x8":       {"internalName": "c3.2xlarge",                        "traits": {"numCores":   8, "totalMemoryMB":   14785, "ephemeralStorageGB":  157}},
-        "mem1_ssd1_x16":      {"internalName": "c3.4xlarge",                        "traits": {"numCores":  16, "totalMemoryMB":   29900, "ephemeralStorageGB":  302}},
-        "mem1_ssd1_x32":      {"internalName": "c3.8xlarge",                        "traits": {"numCores":  32, "totalMemoryMB":   60139, "ephemeralStorageGB":  637}},
-        "mem3_ssd1_x32_gen1": {"internalName": "cr1.8xlarge",                       "traits": {"numCores":  32, "totalMemoryMB":  245751, "ephemeralStorageGB":  237}},
-        "mem3_ssd1_x2":       {"internalName": "r3.large",                          "traits": {"numCores":   2, "totalMemoryMB":   15044, "ephemeralStorageGB":   27}},
-        "mem3_ssd1_x4":       {"internalName": "r3.xlarge",                         "traits": {"numCores":   4, "totalMemoryMB":   30425, "ephemeralStorageGB":   72}},
-        "mem3_ssd1_x8":       {"internalName": "r3.2xlarge",                        "traits": {"numCores":   8, "totalMemoryMB":   61187, "ephemeralStorageGB":  147}},
-        "mem3_ssd1_x16":      {"internalName": "r3.4xlarge",                        "traits": {"numCores":  16, "totalMemoryMB":  122705, "ephemeralStorageGB":  297}},
-        "mem3_ssd1_x32":      {"internalName": "r3.8xlarge",                        "traits": {"numCores":  32, "totalMemoryMB":  245751, "ephemeralStorageGB":  597}}
-}"""
-
-    private val awsOnDemandHourlyPrice =
-                """|{
-           | "cc2.8xlarge": 2.000,
-           | "cg1.4xlarge": 2.100,
-           | "m4.large": 0.108,
-           | "m4.xlarge": 0.215,
-           | "m4.2xlarge": 0.431,
-           | "m4.4xlarge": 0.862,
-           | "m4.10xlarge": 2.155,
-           | "m4.16xlarge": 3.447,
-           | "c4.large": 0.100,
-           | "c4.xlarge": 0.199,
-           | "c4.2xlarge": 0.398,
-           | "c4.4xlarge": 0.796,
-           | "c4.8xlarge": 1.591,
-           | "p2.xlarge": 0.900,
-           | "p2.8xlarge": 7.200,
-           | "p2.16xlarge": 14.400,
-           | "g2.2xlarge": 0.650,
-           | "g2.8xlarge": 2.600,
-           | "x1.16xlarge": 6.669,
-           | "x1.32xlarge": 13.338,
-           | "r4.large": 0.133,
-           | "r4.xlarge": 0.266,
-           | "r4.2xlarge": 0.532,
-           | "r4.4xlarge": 1.064,
-           | "r4.8xlarge": 2.128,
-           | "r4.16xlarge": 4.256,
-           | "r3.large": 0.166,
-           | "r3.xlarge": 0.333,
-           | "r3.2xlarge": 0.665,
-           | "r3.4xlarge": 1.330,
-           | "r3.8xlarge": 2.660,
-           | "i2.xlarge": 0.853,
-           | "i2.2xlarge": 1.705,
-           | "i2.4xlarge": 3.410,
-           | "i2.8xlarge": 6.820,
-           | "d2.xlarge": 0.690,
-           | "d2.2xlarge": 1.380,
-           | "d2.4xlarge": 2.760,
-           | "d2.8xlarge": 5.520,
-           | "hi1.4xlarge": 3.100,
-           | "hs1.8xlarge": 4.600,
-           | "m3.medium": 0.067,
-           | "m3.large": 0.133,
-           | "m3.xlarge": 0.266,
-           | "m3.2xlarge": 0.532,
-           | "c3.large": 0.090,
-           | "c3.xlarge": 0.210,
-           | "c3.2xlarge": 0.420,
-           | "c3.4xlarge": 0.840,
-           | "c3.8xlarge": 1.680,
-           | "m1.small": 0.044,
-           | "m1.medium": 0.087,
-           | "m1.large": 0.175,
-           | "m1.xlarge": 0.350,
-           | "c1.medium": 0.130,
-           | "c1.xlarge": 0.520,
-           | "m2.xlarge": 0.245,
-           | "m2.2xlarge": 0.490,
-           | "m2.4xlarge": 0.980,
-           | "t1.micro": 0.020,
-           | "cr1.8xlarge": 3.500
-           |}
-                   |""".stripMargin.trim
-
-
-    // Create an availble instance list based on a hard coded list
-    def genTestDB(pricingInfo: Boolean) : InstanceTypeDB = {
-        def intOfJs(jsVal : JsValue) : Int = {
-            jsVal match {
-                case JsNumber(x) => x.toInt
-                case _ => throw new Exception("sanity")
-            }
-        }
-        val awsOnDemandHourlyPriceTable: Map[String, Float] = {
-            val fields : Map[String, JsValue] = awsOnDemandHourlyPrice.parseJson.asJsObject.fields
-            fields.map{ case(name, v) =>
-                val price: Float = v match {
-                    case JsNumber(x) => x.toFloat
-                    case _ => throw new Exception("sanity")
-                }
-                name -> price
-            }.toMap
-        }
-
-        val allInstances : Map[String, JsValue] = instanceList.parseJson.asJsObject.fields
-        val db = allInstances.map{ case(name, v) =>
-            val fields : Map[String, JsValue] = v.asJsObject.fields
-            val internalName = fields("internalName") match {
-                case JsString(s) => s
-                case _ => throw new Exception("sanity")
-            }
-            val price: Float =
-                if (pricingInfo) awsOnDemandHourlyPriceTable(internalName)
-                else 0
-            val traits = fields("traits").asJsObject.fields
-            val memoryMB = intOfJs(traits("totalMemoryMB"))
-            val diskGB = intOfJs(traits("ephemeralStorageGB"))
-            val cpu = intOfJs(traits("numCores"))
-            DxInstanceType(name, memoryMB, diskGB, cpu, price, Vector.empty)
-        }.toVector
-        InstanceTypeDB(db)
+        InstanceTypeDB(true, opaque)
     }
 }